requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
lightgbm
joblib
python-dotenv
sqlalchemy
psycopg2-binary
reportlab
